backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0


openai==1.52.2


python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3


nltk==3.9.1
scikit-learn==1.5.2

pdfminer.six==20231228
